Modulo Personas acomodado

+++ database/migrations/2020_02_01_203511_create_compras_criptomoneda_table.php
@@ -29,8 +29,8 @@ class CreateComprasCriptomonedaTable extends Migration
 
             $table->decimal('precio_moneda_a_comprar',13,9);
             $table->decimal('precio_moneda_a_pagar',13,9);
-            $table->tinyInteger('comision_general');
-            $table->tinyInteger('comision_compra');
+            $table->double('comision_general',5,2);
+            $table->double('comision_compra',5,2);
             $table->decimal('monto',13,9);
             $table->decimal('total_sin_comision',13,9);
             $table->decimal('total_con_comision',13,9);
